requirements.txt
PyYAML~=6.0
XlsxWriter~=3.0.4
pandas~=1.5.1
python-dateutil~=2.8.2
requests~=2.28.1
botocore~=1.29.15
urllib3~=1.26.12
python-dotenv~=0.21.0